Added eol harvest test for worker
