+++ tests/functional/worker/eol-harvest.test.js
@@ -1,0 +1,65 @@
+const testDriver = require('../../../tools/jil/index')
+const {workerTypes, typeToMatcher} = require('./helpers')
+const {fail} = require('../uncat-internal-help.cjs')
+
+workerTypes.forEach(type => {  // runs all test for classic & module workers & use the 'workers' browser-matcher for classic and the 'workersFull' for module
+	const browsersWithOrWithoutModuleSupport = typeToMatcher(type);
+	finalHarvest(type, browsersWithOrWithoutModuleSupport);
+});
+
+// --- Tests ---
+function finalHarvest (type, browserVersionMatcher) {
+	testDriver.test(`${type} - buffered events are sent at end-of-life aka worker closing`, browserVersionMatcher, 
+		function (t, browser, router) {
+			// CAUTION: RegExp objs must be explicitly stringified to be transferred, and atm our code reads it as a RegExp rather than a string
+			let assetURL = router.assetURL(`worker/${type}-worker.html`, {
+				init: {
+					ajax: { harvestTimeSeconds: 81 },
+					jserrors: { harvestTimeSeconds: 81 },
+					ins: { harvestTimeSeconds: 81 }
+				},
+				workerCommands: [() => {
+					newrelic.noticeError("test");
+					newrelic.addPageAction("blahblahblah");
+					fetch('/json');
+					self.close();
+				}].map(x => x.toString())
+			});
+
+			const loadPromise = browser.get(assetURL);
+			const metrPromise = router.expectErrors();
+			const errPromise = router.expectErrors();		// CAUTION: the order of metrics (sm) and jserrors matters; metrics are always sent out FIRST
+			const ajaxPromise = router.expectAjaxEvents();
+			const insPromise = router.expectIns();
+
+			Promise.all([loadPromise, metrPromise, errPromise, ajaxPromise, insPromise])
+			.then(( [, metrResponse, errResponse, ajaxResponse, insResponse] ) => {
+				let body;
+
+				body = JSON.parse(metrResponse.body);
+				t.ok(body.sm, 'supportability metrics are sent on close');
+				t.ok(body.sm.length >= 2, 'metrics included api calls as expected');
+				t.equal(metrResponse.req.method, 'POST', 'metrics(-jserror) harvest is a POST');
+
+				body = JSON.parse(errResponse.body);
+				t.ok(body.err, 'jserrors are sent on close');
+				t.equal(body.err.length, 1, 'should have 1 error obj');
+				t.equal(body.err[0].params.message, 'test', 'should have correct message');
+				t.equal(errResponse.req.method, 'POST', 'jserrors harvest is a POST');
+
+				body = ajaxResponse.body;
+				t.ok(/^bel.*\/jserrors.*\/ins.*$/.test(body), 'ajax event is sent on close with right payload');
+				t.equal(errResponse.req.method, 'POST', 'events harvest is a POST');
+
+				body = JSON.parse(insResponse.body);
+				t.ok(body.ins, 'page actions are sent on close');
+				t.equal(body.ins.length, 1, 'should have 1 action obj');
+				t.equal(body.ins[0].actionName, 'blahblahblah', 'should have correct actionName');
+				t.equal(insResponse.req.method, 'POST', 'ins harvest is a POST');
+
+      	t.end()
+			})
+			.catch(fail(t));
+		}
+	);
+}
