--- tests/functional/worker/final-harvest.test.js
@@ -1,5 +1,0 @@
-// TO DO: "PageAction submission on final harvest" test from page-action-submission
-
-// TO DO: do we need to re-evaluate browser matcher reliablefinalharvest and unload?
-
-// TO DO: translate final-harvest.test.js from functional folder
